requirements.txt
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
Django==5.1.2
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
Markdown==3.7
MarkupSafe==3.0.1
mysqlclient==2.2.4
PyJWT==2.9.0
sqlparse==0.5.1
Werkzeug==3.0.4
